moved checkout function to db_carts and cleaned up old functions
