+++ routes/cartsRouter.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const cartsRouter = express.Router();
 const db_carts = require("../db/carts.js");
-const db_orders = require("../db/orders.js");
 
 module.exports = cartsRouter;
 
@@ -13,15 +12,12 @@ cartsRouter.get("/:id", db_carts.getCartById);
 
 //POST cart
 cartsRouter.post("/", db_carts.addProductToCart);
-cartsRouter.post(
-  "/checkout/:id",
-  db_orders.checkoutCart
-);
+cartsRouter.post("/checkout/", db_carts.checkoutCart);
 
 //PUT cart
-cartsRouter.put("/updatecart/:id", db_carts.updateCart)
-cartsRouter.put('/plusone/', db_carts.addOneToCart);
-cartsRouter.put('/minusone/', db_carts.subtractOneFromCart)
+cartsRouter.put("/updatecart/:id", db_carts.updateCart);
+cartsRouter.put("/plusone/", db_carts.addOneToCart);
+cartsRouter.put("/minusone/", db_carts.subtractOneFromCart);
 
 //DELETE cart
 cartsRouter.delete("/:id", db_carts.deleteCart);
